Bring in E3SM cime changes for past 3 months

Major change list:

Python

Minor improvements to bless_test_results.py
Add mach-specific throughput checking to test machines
Some tools were not loading environment before running which was causing problems on some platforms
Improvements in robustness, error msgs of check_input_data
configure.py new FakeCase class
add FakeCase to fortran_unit_testing/run_tests.py
system_tests_common.py change to tput test
env_batch.py add new CHARGE_ACCOUNT
env_mach_specific.py - Allow config_machines.xml to have any case variable as an attribute,
pes.py - also output threads, root
generic_xml.py - remove minidom
build.py - add USE_MOAB option
case.py - use new CHARGE_ACCOUNT. new SMP_PRESENT
provenancy.py - update E3SM provenance.
test_scheduler.py - tput tolerance. new "R" modifier for single column tests.
utils.py new get_charge_account method
update_acme_tests.py - more tests and test suites.
schema changes

config_compilers_v2.xsd: Add MOAB
config_machines.xsd: Add CHARGE_ACCOUNT for charging to a different name then PROJECT used for directory paths
config_machines_template.xml: Also add CHARGE_ACCOUNT
env_mach_specific.xsd: Change debug to upperBoolean
Tools

new acme_cime_split to help with subtree splitting
new tests for the run_acme script in scripts/test/run_acme_test
xml changes

Change "debug" to "DEBUG" in all config_machines.xml
** data models **

datm/cime_config/namelist_definition_datm.xml new 6hourly forcing data set
** mct driver **

config_component.xml new SMP_PRESENT, USE_MOAB, TEST_TPUT_TOLERANCE, CHARGE_ACCOUNT
cime_comp_mod.F90 new option to shr_mem_getusage to also print.
seq_diag_mct.F90 fix an E3SM diag
timing

fix for BGQ
shr_reprosum_mod.F90 - optionally check for nans.

Rename some man3 pages in PIO1

Many E3SM config changes.

Test suite: scripts_regression_tests
Test baseline:
Test namelist changes:
Test status: bit for bit

Fixes ESMCI/cime#1332
Fixes ESMCI/cime#1626

User interface changes?:

Update gh-pages html N

Code review:

+++ scripts/lib/CIME/hist_utils.py
@@ -5,7 +5,7 @@ Functions for actions pertaining to history files.
 from CIME.XML.standard_module_setup import *
 from CIME.test_status import TEST_NO_BASELINES_COMMENT
 
-import logging, glob, os, shutil, re, stat
+import logging, glob, os, shutil, re, stat, filecmp
 logger = logging.getLogger(__name__)
 
 def _iter_model_file_substrs(case):
@@ -212,7 +212,13 @@ def _compare_hists(case, from_dir1, from_dir2, suffix1="", suffix2="", outfile_s
             else:
                 comments += "    {} did NOT match {}\n".format(hist1, hist2)
                 comments += "    cat " + cprnc_log_file + "\n"
-                shutil.copy(cprnc_log_file, casedir)
+                expected_log_file = os.path.join(casedir, os.path.basename(cprnc_log_file))
+                if not (os.path.exists(expected_log_file) and filecmp.cmp(cprnc_log_file, expected_log_file)):
+                    try:
+                        shutil.copy(cprnc_log_file, casedir)
+                    except OSError:
+                        logger.warning("Could not copy {} to {}".format(cprnc_log_file, casedir))
+
                 all_success = False
 
     if num_compared == 0:
